requirements.txt
feedparser
google-genai
python-dotenv
pywin32
